mark as read on window focus
